[ci fw-only Java/greenlightning]
swap versions
