+++ frameworks/Java/greenlightning/src/main/java/com/javanut/gl/benchmark/Field.java
@@ -1,4 +1,4 @@
-package com.ociweb.gl.benchmark;
+package com.javanut.gl.benchmark;
 
 public enum Field {
 	//identifiers for each field used by routes and/or structures
