+++ frameworks/Java/greenlightning/src/main/java/com/javanut/gl/benchmark/FortunesObject.java
@@ -1,4 +1,4 @@
-package com.ociweb.gl.benchmark;
+package com.javanut.gl.benchmark;
 
 import java.util.ArrayList;
 import java.util.Collections;
