+++ frameworks/Java/greenlightning/src/main/java/com/javanut/gl/benchmark/ProcessQuery.java
@@ -1,13 +1,13 @@
-package com.ociweb.gl.benchmark;
+package com.javanut.gl.benchmark;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.ThreadLocalRandom;
 
-import com.ociweb.gl.api.HTTPRequestReader;
-import com.ociweb.gl.api.HTTPResponseService;
-import com.ociweb.pronghorn.network.config.HTTPContentTypeDefaults;
-import com.ociweb.pronghorn.pipe.ObjectPipe;
+import com.javanut.gl.api.HTTPRequestReader;
+import com.javanut.gl.api.HTTPResponseService;
+import com.javanut.pronghorn.network.config.HTTPContentTypeDefaults;
+import com.javanut.pronghorn.pipe.ObjectPipe;
 
 import io.reactiverse.pgclient.PgIterator;
 import io.reactiverse.pgclient.PgPool;
