+++ frameworks/Java/greenlightning/src/main/java/com/javanut/gl/benchmark/ResultObject.java
@@ -1,4 +1,4 @@
-package com.ociweb.gl.benchmark;
+package com.javanut.gl.benchmark;
 
 public class ResultObject {
