+++ frameworks/Java/greenlightning/src/main/java/com/javanut/gl/benchmark/Struct.java
@@ -1,4 +1,4 @@
-package com.ociweb.gl.benchmark;
+package com.javanut.gl.benchmark;
 
 public enum Struct {
 	PLAINTEXT_ROUTE, JSON_ROUTE,
