--- frameworks/Java/greenlightning/src/main/java/com/ociweb/gl/benchmark/GreenLightning.java
@@ -1,17 +1,0 @@
-package com.ociweb.gl.benchmark;
-
-import com.ociweb.gl.api.GreenRuntime;
-import com.ociweb.pronghorn.stage.scheduling.GraphManager;
-
-public class GreenLightning {
-
-	public static void main(String[] args) {
-
-		GraphManager.showThreadIdOnTelemetry = true;
-		GraphManager.showScheduledRateOnTelemetry = true;
-		                      
-		GreenRuntime.run(new FrameworkTest(),args);
-	
-	}
-	
-}
