+++ frameworks/Java/greenlightning/src/main/resources/GreenLight.props
@@ -1,1 +1,1 @@
-main=com.ociweb.gl.benchmark.FrameworkTest
+main=com.javanut.gl.benchmark.FrameworkTest
